[PATCH] Finish test re-organization to prep for functional tests
Simple change that moves integration tests from the `unit` directory to the `integration` directory, and also uses `coverage.py` directly instead of the `pytest-cov` plugin in order to have truly accurate coverage numbers. This is in preparation for adding more integration tests and extensive functional tests.
