+++ .gitignore
@@ -13,7 +13,7 @@ dist/
 
 .idea
 *.iml
-.coverage
+.coverage*
 coverage.xml
 pytests.xml
 .pytest_cache
