+++ .travis.yml
@@ -14,21 +14,31 @@ install:
 jobs:
    include:
      - stage: build
+       name: "Unit Tests: Python 2.7 with PyTest 3.1 and 4.6"
        python: "2.7"
        script:
-         - tox -e py27,py27-pytest31,py27-pytest46,py27-flake8,coverage
+         - tox -e py27-pytest31,py27-pytest46,py27-flake8,coverage
      - stage: build
+       name: "Unit Tests: Python 3.5 with and without PyInotify"
        python: "3.5"
        script:
          - tox -e py35,py35-pyinotify,coverage
      - stage: build
+       name: "Unit Tests: Python 3.6 with Currint 1.6 and 2.0"
        python: "3.6"
        script:
          - tox -e py36,py36-currint16,py36-currint20,coverage
      - stage: build
+       name: "Unit Tests: Python 3.7 with Attrs 17, 18, and 19"
        python: "3.7"
        script:
-         - tox -e py37,py37-attrs17,py37-attrs18,py37-attrs19,py37-pytest31,py37-pytest46,py37-pytest50,py37-flake8,coverage
+         - tox -e py37-attrs17,py37-attrs18,py37-attrs19,py37-flake8,coverage
+       dist: xenial
+     - stage: build
+       name: "Unit Tests: Python 3.7 with PyTest 3.1, 4.6, and 5.0"
+       python: "3.7"
+       script:
+         - tox -e py37-pytest31,py37-pytest46,py37-pytest50,py37-flake8,coverage
        dist: xenial
      - stage: deploy
        python: "3.6"
