+++ setup.cfg
@@ -32,8 +32,17 @@ known_third_party=attr,conformity,currint,dateutil,ddtrace,django,factory,faker,
 known_current_project=pysoa
 known_tests=tests
 
+[coverage:run]
+branch = true
+source = pysoa
+
+[coverage:report]
+fail_under = 85
+precision = 1
+show_missing = true
+
 [aliases]
-test=pytest
+test = pytest
 
 [tool:pytest]
 addopts = -s
