+++ test.sh
@@ -1,5 +1,22 @@
-#!/bin/bash -ex
+#!/bin/bash
 
 # For running tests locally (not Docker, not Tox) with all dependencies already installed
 
-pytest --cov=pysoa --cov-branch --cov-report=term-missing --cov-fail-under=85
+coverage run --parallel-mode -m pytest tests/unit tests/integration
+RET=$?
+
+coverage combine
+ret_sub=$?
+if [[ $ret_sub -gt 0 ]] && [[ $RET -eq 0 ]]
+then
+    RET=$ret_sub
+fi
+
+coverage report
+ret_sub=$?
+if [[ $ret_sub -gt 0 ]] && [[ $RET -eq 0 ]]
+then
+    RET=$ret_sub
+fi
+
+exit $RET
