+++ tests/functional/test_fail.py
@@ -1,0 +1,5 @@
+from __future__ import absolute_import, unicode_literals
+
+
+def test_fail():
+    raise AssertionError('This test should not be run yet')
