+++ tests/integration/test/plan/first_fixtures/003_globals.pysoa
