+++ tests/integration/test/plan/mocking_and_stubbing/stubbing_test.pysoa
