+++ tests/integration/test/plan/skipped_fixtures/global_skip.pysoa
