+++ tests/integration/test/plan/test_002_plugin_stats.py
@@ -9,8 +9,7 @@ import pytest
 
 from pysoa.test.plugins.pytest.plans import PLUGIN_STATISTICS
 
-from tests.unit.test.plan import test_001_fixtures_work as fixtures_test_module
-
+from tests.integration.test.plan import test_001_fixtures_work as fixtures_test_module
 
 _skip_correction = 0
 if LooseVersion(pytest.__version__) > LooseVersion('4.2.0'):
