+++ tests/integration/test/plan/tiny-image.gif
