+++ tox.ini
@@ -29,7 +29,7 @@ deps =
 commands =
 #    conformity_branch: pip uninstall -y conformity
 #    conformity_branch: pip install git+https://github.com/eventbrite/conformity.git@insert_branch_name_here
-    pytest --cov=pysoa --cov-append --cov-branch --cov-report=
+    coverage run --parallel-mode -m pytest tests/unit tests/integration
 
 [testenv:py27-flake8]
 skip_install = true
@@ -44,4 +44,6 @@ commands = flake8
 [testenv:coverage]
 skip_install = true
 deps = coverage
-commands = coverage report -m --fail-under 85
+commands =
+    coverage combine
+    coverage report
